chore: migrate ProvidersOverrideTest to a new test framework
closes: https://github.com/keycloak/keycloak/issues/50773
