+++ tests/custom-providers/src/main/java/org/keycloak/tests/providers/providersoverride/CustomDefaultEmailSenderProvider2.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.keycloak.examples.providersoverride;
+package org.keycloak.tests.providers.providersoverride;
 
 import java.util.Map;
