+++ tests/custom-providers/src/main/java/org/keycloak/tests/providers/providersoverride/CustomLoginFormsProvider.java
@@ -17,24 +17,17 @@
  *
  */
 
-package org.keycloak.examples.providersoverride;
+package org.keycloak.tests.providers.providersoverride;
 
-import org.keycloak.email.DefaultEmailSenderProviderFactory;
-import org.keycloak.email.EmailSenderProvider;
+import org.keycloak.forms.login.freemarker.FreeMarkerLoginFormsProvider;
 import org.keycloak.models.KeycloakSession;
 
 /**
- * Test for order (This one should be called in favour of FreemarkerAccountProviderFactory and CustomFreemarkerAccountProviderFactory1 as it has highest order)
+ * @author <a href="mailto:mposolda@redhat.com">Marek Posolda</a>
  */
-public class CustomDefaultEmailSenderProviderFactory2 extends DefaultEmailSenderProviderFactory {
+public class CustomLoginFormsProvider extends FreeMarkerLoginFormsProvider {
 
-    @Override
-    public int order() {
-        return 2;
-    }
-
-    @Override
-    public EmailSenderProvider create(KeycloakSession session) {
-        return new CustomDefaultEmailSenderProvider2(session, getEmailAuthenticators());
+    public CustomLoginFormsProvider(KeycloakSession session) {
+        super(session);
     }
 }
