+++ tests/custom-providers/src/main/java/org/keycloak/tests/providers/providersoverride/CustomLoginFormsProviderFactory.java
@@ -17,17 +17,22 @@
  *
  */
 
-package org.keycloak.examples.providersoverride;
+package org.keycloak.tests.providers.providersoverride;
 
-import org.keycloak.forms.login.freemarker.FreeMarkerLoginFormsProvider;
+import org.keycloak.forms.login.LoginFormsProvider;
+import org.keycloak.forms.login.freemarker.FreeMarkerLoginFormsProviderFactory;
 import org.keycloak.models.KeycloakSession;
 
 /**
- * @author <a href="mailto:mposolda@redhat.com">Marek Posolda</a>
+ * This has same providerID like built-in FreeMarkerLoginFormsProviderFactory. But it should be called in favour of it even
+ * if it doesn't have "order" set. As it is custom provider and it worked this way in previous versions
  */
-public class CustomLoginFormsProvider extends FreeMarkerLoginFormsProvider {
+public class CustomLoginFormsProviderFactory extends FreeMarkerLoginFormsProviderFactory {
 
-    public CustomLoginFormsProvider(KeycloakSession session) {
-        super(session);
+    @Override
+    public LoginFormsProvider create(KeycloakSession session) {
+        return new CustomLoginFormsProvider(session);
     }
+
+
 }
