+++ tests/custom-providers/src/main/java/org/keycloak/tests/providers/providersoverride/CustomValidateUsername.java
@@ -17,22 +17,14 @@
  *
  */
 
-package org.keycloak.examples.providersoverride;
+package org.keycloak.tests.providers.providersoverride;
 
-import org.keycloak.forms.login.LoginFormsProvider;
-import org.keycloak.forms.login.freemarker.FreeMarkerLoginFormsProviderFactory;
-import org.keycloak.models.KeycloakSession;
+import org.keycloak.authentication.authenticators.directgrant.ValidateUsername;
 
 /**
  * This has same providerID like built-in ValidateUsername provider. But it should be called in favour of ValidateUsername even
  * if it doesn't have "order" set. As it is custom provider and it worked this way in previous versions
+ *
  */
-public class CustomLoginFormsProviderFactory extends FreeMarkerLoginFormsProviderFactory {
-
-    @Override
-    public LoginFormsProvider create(KeycloakSession session) {
-        return new CustomLoginFormsProvider(session);
-    }
-
-
+public class CustomValidateUsername extends ValidateUsername {
 }
