+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.authentication.AuthenticatorFactory
@@ -1,1 +1,6 @@
 org.keycloak.tests.providers.forms.ClickThroughAuthenticator
+org.keycloak.tests.providers.providersoverride.CustomValidateUsername
+org.keycloak.tests.providers.providersoverride.CustomValidatePassword1
+org.keycloak.tests.providers.providersoverride.CustomValidatePassword2
+org.keycloak.tests.providers.providersoverride.CustomValidatePassword3
+org.keycloak.tests.providers.providersoverride.CustomValidateOTP
