+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.email.EmailSenderProviderFactory
@@ -1,0 +1,2 @@
+org.keycloak.tests.providers.providersoverride.CustomDefaultEmailSenderProviderFactory1
+org.keycloak.tests.providers.providersoverride.CustomDefaultEmailSenderProviderFactory2
