+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.forms.login.LoginFormsProviderFactory
@@ -1,0 +1,1 @@
+org.keycloak.tests.providers.providersoverride.CustomLoginFormsProviderFactory
