+++ testsuite/integration-arquillian/tests/base/testsuites/base-suite
@@ -19,7 +19,6 @@ oauth,6
 oidc,6
 organization,3
 policy,6
-providers,4
 saml,6
 script,6
 theme,6
